Moved admin courses to general
