--- app/views/admin/course_names/edit.html.erb
@@ -1,23 +1,0 @@
-<section>
-  <h2 class="text-center fw-bold py-3">Editing <b><%= @course.name.capitalize %></b> Course</h2>
-  <br />
-  <br />
-    <%= form_for @course, url: { controller: 'admin/course_names', action: 'update' } do |f| %>
-      <div class="container">
-        <div class="row text-center">
-          <div class="col-md-12 form-group input-group">
-            <span class="input-group-text">Course Name:</span>
-            <%= f.text_field :name, placeholder: 'Write an unique name...', class: 'form-control', required: true %>
-            <%= f.submit 'Update Course', class: 'btn btn-primary btn-xl' %>
-            <%= link_to 'Back', admin_course_names_path, class: 'btn btn-secondary' %>
-            <%= button_to 'Delete', { controller: 'admin/course_names', action: :destroy, id: @course.id }, method: :delete,
-                                                                                                            data: { confirm: 'Are you sure that you want to delete this course ?' },
-                                                                                                            class: 'btn btn-danger' %>
-          </div>
-        </div>
-      </div>
-    <% end %>
-
-
-  <br /><br />
-</section>
