--- app/views/admin/course_names/index.html.erb
@@ -1,38 +1,0 @@
-<section>
-    <h2 class="text-center fw-bold py-3">Admin Course Manager</h2>
-
-    <%= form_with url: {controller: "admin/course_names", action: "create"} do |f| %>
-      <div class="container">
-        <div class="row text-center">
-          <div class="col-auto form-group input-group">
-            <span class="input-group-text">Course Name</span>
-            <%= f.text_field :name, placeholder: "Write an unique name...", class: 'form-control', required: true %>
-            <%= f.submit "Create Course", class: "btn btn-primary btn-xl" %>
-            </div>
-        </div>
-      </div>
-    <% end %>
-    <br />
-   <table class="table table-striped mb-3">
-   <thead class="table-primary">
-      <tr class="text-center">
-        <th>Course</th>
-        <th>Action</th>
-      </tr>
-    </thead>
-      <% @courses.each do |t| %>
-        <tr>
-          <td><%= t.name %></td>
-          <td>
-          <div>
-            <%= link_to "Edit", edit_admin_course_name_path(t.id), class:"btn btn-primary" %>
-            <%= button_to "Delete", admin_course_name_path(t.id), method: :delete, data: { confirm: "Are you sure?" }, class:"btn btn-danger", form_class:"my-2" %>
-          </div>
-          </td>
-
-        </tr>
-      <% end %>
-
-    </table>
-  <br>
-</section>
