--- app/views/admin/course_names/new.html.erb
@@ -1,21 +1,0 @@
-<section>
-  <h2 class="text-center fw-bold py-3">New Course</h2>
-  <div class="float-end">
-    <%= link_to 'Back', admin_course_names_path %>
-  </div>
-  <br />
-    <%= form_for @new_course, url: { controller: 'admin/course_names', action: 'create' } do |f| %>
-      <div class="container">
-        <div class="row text-center">
-          <div class="col-md-12 form-group">
-            Name: <%= f.text_field :name, placeholder: 'Write an unique name...', class: 'form-control', required: true %>
-          </div>
-          <div class="col-md-12 form-group my-1">
-            <%= f.submit 'Create Course', class: 'btn btn-secondary btn-xl text-uppercase' %>
-          </div>
-        </div>
-      </div>
-    <% end %>
-
-  <br><br>
-</section>
