+++ app/views/admin/index.html.erb
@@ -11,7 +11,6 @@
     <li><%= link_to 'Skill Types', admin_skills_path, class: 'list-group-item list-group-item-action' %></li>
     <li><%= link_to 'Popular Hours', admin_report_generator_popular_hours_path, class: 'list-group-item list-group-item-action' %></li>
     <li><%= link_to 'Manage Trainings', admin_trainings_path, class: 'list-group-item list-group-item-action' %></li>
-    <li><%= link_to 'Manage Courses', admin_course_names_path, class: 'list-group-item list-group-item-action' %></li>
     <li><%= link_to 'Training Sessions', admin_training_sessions_path, class: 'list-group-item list-group-item-action' %></li>
     <li><%= link_to 'Manage Spaces', admin_spaces_path, class: 'list-group-item list-group-item-action' %></li>
     <li><%= link_to 'Manage Roles', manage_roles_admin_users_path, class: 'list-group-item list-group-item-action' %></li>
