+++ app/views/layouts/_admin_area_header.html.erb
@@ -12,7 +12,6 @@
         <li class="nav-item"><%= link_to 'Reports', admin_report_generator_index_path, class: 'nav-link' %></li>
         <li class="nav-item"><%= link_to 'Popular Hours', admin_report_generator_popular_hours_path, class: 'nav-link' %></li>
         <li class="nav-item"><%= link_to 'Trainings', admin_trainings_path, class: 'nav-link' %></li>
-        <li class="nav-item"><%= link_to 'Courses', admin_course_names_path, class: 'nav-link' %></li>
         <li class="nav-item"><%= link_to 'Training Sessions', admin_training_sessions_path, class: 'nav-link' %></li>
         <li class="nav-item"><%= link_to 'Spaces', admin_spaces_path, class: 'nav-link' %></li>
         <li class="nav-item dropdown">
